requirements.txt
asgiref==3.11.1
Django==6.0.3
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==
django-cors-headers==4.9.0
